feat: Metadata: 租户初始化时将kafka集群注册到gse --story=127096638 (#7985)
