+++ bkmonitor/metadata/task/tenant.py
@@ -9,6 +9,7 @@ from metadata.models import ClusterInfo
 from metadata.task.bkbase import sync_bkbase_cluster_info
 from metadata.task.constants import BKBASE_V4_KIND_STORAGE_CONFIGS
 from metadata.task.sync_space import sync_bkcc_space
+from metadata.utils.gse import KafkaGseSyncer
 
 logger = logging.getLogger(__name__)
 
@@ -21,25 +22,31 @@ def _init_kafka_cluster(bk_tenant_id: str):
 
     # 如果有默认的kafka集群，则不再初始化
     if kafka_cluster_query.filter(bk_tenant_id=bk_tenant_id, is_default_cluster=True).exists():
+        cluster = kafka_cluster_query.get(bk_tenant_id=bk_tenant_id, is_default_cluster=True)
         logger.info("Kafka cluster already exists for tenant %s, skipping initialization.", bk_tenant_id)
-        return
+    else:
+        # 创建默认的kafka集群
+        kafka_host = os.environ.get("BK_MONITOR_KAFKA_HOST", None)
+        kafka_port = os.environ.get("BK_MONITOR_KAFKA_PORT", None)
+        if not kafka_host or not kafka_port:
+            logger.error("BK_MONITOR_KAFKA_HOST and BK_MONITOR_KAFKA_PORT must be set in environment variables")
+            raise ValueError("BK_MONITOR_KAFKA_HOST and BK_MONITOR_KAFKA_PORT must be set in environment variables")
 
-    # 创建默认的kafka集群
-    kafka_host = os.environ.get("BK_MONITOR_KAFKA_HOST", None)
-    kafka_port = os.environ.get("BK_MONITOR_KAFKA_PORT", None)
-    if not kafka_host or not kafka_port:
-        logger.error("BK_MONITOR_KAFKA_HOST and BK_MONITOR_KAFKA_PORT must be set in environment variables")
-        raise ValueError("BK_MONITOR_KAFKA_HOST and BK_MONITOR_KAFKA_PORT must be set in environment variables")
+        cluster = ClusterInfo.objects.create(
+            bk_tenant_id=bk_tenant_id,
+            cluster_type=ClusterInfo.TYPE_KAFKA,
+            cluster_name="kafka_cluster1",
+            domain_name=kafka_host,
+            port=kafka_port,
+            is_default_cluster=True,
+        )
 
-    ClusterInfo.objects.create(
-        bk_tenant_id=bk_tenant_id,
-        cluster_type=ClusterInfo.TYPE_KAFKA,
-        cluster_name="kafka_cluster1",
-        domain_name=kafka_host,
-        port=kafka_port,
-        is_default_cluster=True,
-    )
-    logger.info("Kafka cluster created for tenant %s", bk_tenant_id)
+        logger.info("Kafka cluster created for tenant %s", bk_tenant_id)
+
+    # 注册到GSE
+    KafkaGseSyncer.register_to_gse(mq_cluster=cluster)
+
+    logger.info("Kafka cluster registered to GSE for tenant %s", bk_tenant_id)
 
 
 def _init_es_cluster(bk_tenant_id: str):
